+++ optd-datafusion-repr/src/cost/stats.rs
@@ -121,13 +121,13 @@ impl<M: MostCommonValues, D: Distribution> PerColumnStats<M, D> {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PerTableStats<M: MostCommonValues, D: Distribution> {
-    row_cnt: usize,
+    pub row_cnt: usize,
     // This is a Vec of Options instead of just a Vec because some columns may not have stats
     //   due to their type being non-comparable.
     // Further, I chose to represent it as a Vec of Options instead of a HashMap because a Vec
     //   of Options clearly differentiates between two different failure modes: "out-of-bounds
     //   access" and "column has no stats".
-    per_column_stats_vec: Vec<Option<PerColumnStats<M, D>>>,
+    pub per_column_stats_vec: Vec<Option<PerColumnStats<M, D>>>,
 }
 
 impl<M: MostCommonValues, D: Distribution> PerTableStats<M, D> {
